requirements.txt
Django>=5.0,<5.1
psycopg2-binary
django-unfold